fix: moved duration to natural language to format
so that it can be reused in course detail
